fix warning, add new image no background

+++ 05-SceneManager/Koopas.h
@@ -58,7 +58,7 @@ protected:
 	float ay;
 
 
-	DWORD defend_start;
+	ULONGLONG defend_start;
 
 	BOOLEAN isDefend;
 	BOOLEAN isKicked;
